% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_cor.R
\name{num_cor}
\alias{num_cor}
\title{A correlation Function}
\usage{
num_cor(year)
}
\arguments{
\item{love}{}
}
\description{
This function creates a correlation table of all numeric variables in a given year
}
\examples{
num_cor()
}
\keyword{nba}
